requirements.txt
numpy==1.18.1
Pillow==6.2.0
six==1.14.0
torch==1.3.1+cu92
torchvision==0.4.2+cu92
